requirements.txt
streamlit==1.41.1
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
matplotlib==3.9.2
torch>=2.1.0
ultralytics==8.3.0
git+https://github.com/facebookresearch/segment-anything.git
